change file name and folder and refactor structure
