+++ modules/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Account"]
+
+from account import Account
