+++ modules/account/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Account"]
+
+from .account_module import Account
