+++ modules/account/account_module.py
@@ -1,24 +1,17 @@
-from asyncio import sleep
-from typing import Dict, Literal, TypedDict
+from typing import List
 
-from modules.client import FTMExplorerClient
-from modules.custom_logger import logger as logging
+from client import FTMExplorerClient
+from utils.custom_logger import logger as logging
 
-ModuleType = Literal["account", "transaction", "contract"]
-ActionType = Literal["balance"]
-
-
-class RequesePayloadInterface(TypedDict):
-    module: ModuleType
-    action: ActionType
-    address: str
+from .account_typing import GetFtmBalanceForSingleAddressInterface
 
 
 class Account(FTMExplorerClient):
 
     async def get_ftm_balance_for_single_address(self, address: str):
+
         logging.info(f"try to get ftm balance for {address}")
-        request_payload: RequesePayloadInterface = {
+        request_payload: GetFtmBalanceForSingleAddressInterface = {
             "module": "account",
             "action": "balance",
             "address": address,
@@ -26,6 +19,9 @@ class Account(FTMExplorerClient):
         response = await self.send_request(request_payload=request_payload)
         logging.debug(f"Accont ftm balance for {address} is {response}")
         return response
+
+    async def get_ftm_balance_for_multiple_addresses(self, addresses: List[str]):
+        logging.info(f"try to get ftm balances for {addresses}")
 
     def get_token_balance(self, token_contract: str, address: str):
         """
