+++ modules/account/account_typing.py
@@ -1,0 +1,13 @@
+from typing import Literal, TypedDict
+
+ModuleType = Literal["account", "transaction", "contract"]
+ActionType = Literal["balance"]
+
+
+class BaseFtmPayloadInterface(TypedDict):
+    module: ModuleType
+    action: ActionType
+
+
+class GetFtmBalanceForSingleAddressInterface(BaseFtmPayloadInterface):
+    address: str
