+++ utils/__init__.py
@@ -1,0 +1,2 @@
+__all__ = ["logger"]
+from .custom_logger import logger
